% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tejapiconfig.R
\name{Tejapi.api_key}
\alias{Tejapi.api_key}
\title{Query or set Tejapi API key}
\usage{
Tejapi.api_key(api_key)
}
\arguments{
\item{api_key}{Optionally passed parameter to set Tejapi \code{api_key}.}
}
\value{
Returns invisibly the currently set \code{api_key}.
}
\description{
Query or set Tejapi API key
}
\examples{
\dontrun{
Tejapi.api_key('foobar')
}
}
